Restructured sample files and function return.
The sample files are given now directly as job argument, no more "if" in
the code. An exception is thrown if the sample file requested can not be
imported. 

To shorten the names, separate folders are made for cmgTuples sample
files and for post-processed sample files. The code looks only in the
corresponding folder to import the sample file.

Function return done via named tuples, to avoid long lists of return
variables and to avoid bugs in return caused by incorect order of
arguments.