Script to find failed jobs on slurm, producing a new submit file.
It identifies jobs which are submitted but do not have an .out file from
the batch system and jobs which did not finished successfully. To
identify successful jobs, the user must write at the end of his/her
program a message, e.g. "End of my job" 

For failed jobs, it creates a new submit file. Clean up of failed jobs
is not (yet?) done.